requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.8
black==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dill==0.3.9
fastapi==0.115.11
Flask==3.1.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-access-context-manager==0.2.1
google-cloud-appengine-logging==1.6.1
google-cloud-asset==3.29.1
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-storage==2.29.1
google-cloud-core==2.4.1
google-cloud-datacatalog==3.24.1
google-cloud-dataplex==2.8.0
google-cloud-logging==3.11.4
google-cloud-org-policy==1.13.0
google-cloud-os-config==1.20.0
google-cloud-quotas==0.1.17
google-cloud-resource-manager==1.14.1
google-cloud-tasks==2.18.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
opentelemetry-api==1.31.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pylint==3.3.4
pyparsing==3.2.1
pytest==8.3.5
pytest-env==1.1.5
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
